[#162382720 Chore] Restructure server folder
