+++ package.json
@@ -4,12 +4,12 @@
   "description": "RESTful api to Power Fast-Food-Fast Application",
   "main": "server.js",
   "scripts": {
-    "start": "npm run build && node ./build/server.js",
+    "start": "npm run build && node ./build/server/server.js",
     "test": "nyc mocha --require babel-register server/tests/test.js --exit",
     "test-watch": "nodemon --exec  \"npm test\"",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
-    "build": "babel server.js --out-dir build",
-    "dev-start": "babel-watch server.js"
+    "build": "babel server --out-dir build",
+    "dev-start": "babel-watch server/server.js"
   },
   "engines": {
     "node": "8.11.3",
